requirements.txt
# Developed with 9.2.0
Pillow